Beta-into hiru

+++ src/components/ManageUsers/Users.js
@@ -1,17 +1,21 @@
 import React, { useState, useEffect } from "react";
+import { useNavigate } from "react-router-dom";
 import DataTable from "react-data-table-component";
 import { Link } from "react-router-dom";
 import axios from "axios";
 import { Modal, message, Space } from "antd";
 import DefaultHandle from "../DefaultHandle";
+import "../../styles/accessmodal.css";
+import swal from 'sweetalert'; 
 
 function Users() {
-  // State to store the list of users
   const [users, setUsers] = useState([]);
   const [filterUser, setFilterUser] = useState([]);
   const [loading, setLoading] = useState(false);
+  const [isHovered, setIsHovered] = useState(false);
+  const navigate = useNavigate();
+  const [isAccessDeniedVisible, setIsAccessDeniedVisible] = useState(false);
 
-  //Effect hook to fetch users data when the component mounts
   useEffect(() => {
     const loadUsers = async () => {
       setLoading(true);
@@ -19,37 +23,41 @@ function Users() {
         const response = await axios.get("http://localhost:8000/user/all", {
           withCredentials: true,
         });
-        console.log(response.data);
-
-        // Set the users state with the data from the response
         setUsers(response.data.data);
-        //console.log(response.data);
         setFilterUser(response.data.data);
-        console.log(response.data);
       } catch (error) {
-        console.error("Error fetching users:", error);
+        if (error.response && error.response.status === 403) {
+          setIsAccessDeniedVisible(true);
+        } else {
+          console.error("Error fetching users:", error);
+          swal({
+            title: "Error",
+            text: "An error occurred while fetching users.",
+            icon: "error",
+            button: "OK",
+          });
+        }
       } finally {
         setLoading(false);
       }
     };
 
     loadUsers();
-  }, []);
-
-  if (loading) {
-    return <div>Loading...</div>; // Show a loading indicator while data is being fetched
-  }
-  // Function to handle user deletion
+  }, [navigate]);
 
   const handleDelete = async (_id) => {
     try {
+      const userToDelete = users.find(user => user._id === _id);
+      if (userToDelete.role === "admin") {
+        message.error("Cannot delete admin user!");
+        return;
+      }
       await axios.delete(`http://localhost:8000/user/delete/${_id}`, {
         withCredentials: true,
       });
-      // Filter out the deleted user from the users state
-      const updatedUsers = users.filter((user) => user._id !== _id);
+      const updatedUsers = users.filter(user => user._id !== _id);
       setUsers(updatedUsers);
-      setFilterUser(updatedUsers); // Update filterUser state as well
+      setFilterUser(updatedUsers);
       message.success("User deleted successfully!");
     } catch (error) {
       console.error("Error deleting user:", error);
@@ -57,7 +65,6 @@ function Users() {
     }
   };
 
-  // Function to display a confirmation modal before deleting a user
   const showDeleteConfirmation = (_id) => {
     Modal.confirm({
       title: "Confirm Delete",
@@ -66,48 +73,48 @@ function Users() {
       okType: "danger",
       cancelText: "No",
       onOk() {
-        // Call handleDelete function if user confirms deletion
         handleDelete(_id);
       },
     });
   };
 
-  // Search users
   const filterUsers = (event) => {
     const searchValue = event.target.value.toLowerCase();
-
-    const userData = users.filter(
-      (row) =>
-        row.firstName.toLowerCase().includes(searchValue) ||
-        row.userId.toLowerCase().includes(searchValue)
+    const userData = users.filter(row =>
+      row.firstName.toLowerCase().includes(searchValue) ||
+      row.userId.toLowerCase().includes(searchValue)
     );
-
-    setFilterUser(userData); // Update filterUser state with filtered data
+    setFilterUser(userData);
   };
 
   const columns = [
     {
       name: "User Id",
-      selector: (row) => row.userId,
+      selector: row => row.userId,
       sortable: true,
     },
     {
       name: "First Name",
-      selector: (row) => row.firstName,
+      selector: row => row.firstName,
+      sortable: true,
+    },
+    {
+      name: "Last Name",
+      selector: row => row.lastName,
       sortable: true,
     },
     {
       name: "Phone Number",
-      selector: (row) => row.phoneNumber,
+      selector: row => row.phoneNumber,
     },
     {
       name: "User Role",
-      selector: (row) => row.role,
+      selector: row => row.role,
       sortable: true,
     },
     {
       name: "Actions",
-      cell: (row) => (
+      cell: row => (
         <div>
           <Link to={`/admin/userform/update/${row._id}`}>Edit</Link>
           <span style={{ margin: "0 8px" }}>|</span>
@@ -117,43 +124,80 @@ function Users() {
     },
   ];
 
+  const closeModal = () => {
+    setIsAccessDeniedVisible(false);
+    navigate("/admin/dashboard");
+  };
+
+  if (loading) {
+    return <div>Loading...</div>;
+  }
+
   return (
     <DefaultHandle>
       <div style={{ marginBottom: "10px" }}>
-      <div style={{  display: "flex", justifyContent: "space-between", alignItems: "center" }}>
-        <input
-          style={{ marginBottom: "12px", width: "200px" }}
-          type="text end"
-          className="input"
-          placeholder="Search user ..."
-          onChange={filterUsers}
+        <div
+          style={{
+            display: "flex",
+            justifyContent: "space-between",
+            alignItems: "center",
+          }}
+        >
+          <input
+            type="text"
+            placeholder="Search User ..."
+            onChange={filterUsers}
+            onMouseEnter={() => setIsHovered(true)}
+            onMouseLeave={() => setIsHovered(false)}
+            style={{
+              marginBottom: "12px",
+              width: "300px",
+              padding: "5px",
+              border: isHovered ? "1px solid black" : "1px solid #ccc",
+              borderRadius: "5px",
+              transition: "border-color 0.3s",
+            }}
+          />
+        </div>
+
+        <div
+          style={{
+            display: "flex",
+            justifyContent: "flex-end",
+            alignItems: "center",
+            marginBottom: "25px",
+          }}
+        >
+          <Space size={12}></Space>
+          <Link to={"/admin/userform"} style={{ fontSize: "14px" }}>
+            Add User
+          </Link>
+        </div>
+
+        <DataTable
+          columns={columns}
+          data={filterUser}
+          selectableRows
+          fixedHeader
+          pagination
         />
       </div>
-
-      <div
-        style={{
-          display: "flex",
-          justifyContent: "flex-end",
-          alignItems: "center",
-          marginBottom: "25px",
-        }}
+      <Modal
+        title="Access Denied!"
+        visible={isAccessDeniedVisible}
+        onCancel={closeModal}
+        footer={[
+          <button onClick={closeModal} key="back">
+            OK
+          </button>,
+        ]}
       >
-        <Space size={12}></Space>
-        <Link to={"/admin/userform"} style={{ fontSize: "16px" }}>
-          Add User
-        </Link>
-      </div>
-
-      <DataTable
-        columns={columns}
-        data={filterUser}
-        selectableRows
-        fixedHeader
-        pagination
-      />
-      </div>
+        <p>You do not have permission to view this page.</p>
+      </Modal>
     </DefaultHandle>
   );
 }
 
 export default Users;
+
+
